feat: actualización de BD y preparación de entorno Spring Boot

+++ src/main/java/com/studysync/repository/UsuarioRepository.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.studysync.model.Usuario;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
-
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
     Usuario findByEmail(String email);
 }
